Update script-header.html (#948)
A change was made previously for searches from the homepage to match a specified project OR one of the latest specified versions. (Previously, content that only appeared under the "tools" project, for example, and was not version specific, was not included in searches.) However, because we were then always including "corda" project files, versions other than the latest specified versions were also included. I have now removed the "corda" project and some other unnecessary projects.